refactor: add LinkItem to BurgerMenu.tsx
