+++ src/components/BurgerMenu/BurgerMenu.tsx
@@ -1,15 +1,24 @@
 import cn from 'classnames';
 import s from './BurgerMenu.module.scss';
 import cross from '../../assets/image/Header/cross.svg';
+import {NavBarLinks} from '../Header/NavBarLinks';
+import {LinkItem} from '../LinkItem/LinkItem';
 
 type BurgerMenuPropsType = {
     isOpen: boolean,
     setIsOpen: VoidFunction
+    activeLinkHandler: (path: string) => void
+    currentURL: string
 }
 
-export const BurgerMenu = ({isOpen, setIsOpen}: BurgerMenuPropsType) => {
+export const BurgerMenu = ({
+                               isOpen,
+                               setIsOpen,
+                               activeLinkHandler,
+                               currentURL
+                           }: BurgerMenuPropsType) => {
 
-    const burgerMenuHandler = ()=>{
+    const burgerMenuHandler = () => {
         setIsOpen()
     }
 
@@ -22,11 +31,13 @@ export const BurgerMenu = ({isOpen, setIsOpen}: BurgerMenuPropsType) => {
                 <img className={s.burger_open} src={cross} alt={'cross-menu'}
                      onClick={burgerMenuHandler}/>
                 <div className={s.menu_list}>
-                    <span>Наши продукты</span>
-                    <span>О нас</span>
-                    <span>Связатся с нами</span>
-                    <span>Делаем</span>
-                    <span>Сотрудничество</span>
+                    {
+                        NavBarLinks.map((item) => {
+                            return <LinkItem item={item} currentURL={currentURL}
+                                             key={item.path}
+                                             onActivateLink={activeLinkHandler}/>
+                        })
+                    }
                 </div>
             </div>
         </div>
