+++ src/components/LinkItem/LinkItem.module.scss
@@ -181,3 +181,33 @@
     }
   }
 }
+
+@media screen and (min-width: 320px) and (max-width: 768px) {
+
+  .link {
+    text-decoration: none;
+
+    .linkText {
+      @include RobotoRegular(14px, normal);
+      position: relative;
+      background-image: linear-gradient(to right, rgba(255, 0, 61, 0.60) 0%, rgba(5, 0, 227, 0.60) 50%, #5200FF 100%);
+      background-size: 200% 100%;
+      background-position: -100%;
+      -webkit-background-clip: text;
+      -webkit-text-fill-color: transparent;
+      transition: all 0.3s ease-in-out;
+
+      &:hover {
+        background-position: 0;
+        -webkit-background-clip: text;
+        -webkit-text-fill-color: transparent;
+      }
+
+      &_active {
+        background: linear-gradient(91deg, rgba(255, 0, 61, 0.60) 0%, rgba(5, 0, 227, 0.60) 50%, #5200FF 100%);
+        -webkit-background-clip: text;
+        -webkit-text-fill-color: transparent;
+      }
+    }
+  }
+}
